GH-8 - renamed: TemplateEditorInstance.js ->ChelistTemplate.js TemplateEditorInstance.css ->ChelistTemplate.css TemplateEditorInstanceTools.js ->ChelistTemplateTools.js Item.js -> TemplateEditorInstanceItem.js
