+++ src/ChecklistDojo/ClientApp/src/App.jsx
@@ -6,7 +6,7 @@ import LoadingIndicator from "./components/LoadingIndicator";
 import ChecklistHistory from "./components/ChecklistHistory";
 import Settings from "./components/Settings";
 import ChecklistInstance from "./components/ChecklistInstance";
-import TemplateEditorInstance from "./components/TemplateEditorInstance";
+import ChecklistTemplate from "./components/ChecklistTemplate";
 import Templates from "./components/Templates";
 import Scheduled from "./components/Scheduled";
 
@@ -22,7 +22,7 @@ class App extends Component {
           component={props => <Home auth={this.props.auth} {...props} />}
         />
         <Route path="/Checklist/:guid" component={ChecklistInstance} />
-        <Route path="/TemplateEditor/:guid" component={TemplateEditorInstance} />
+        <Route path="/ChecklistTemplate/:guid" component={ChecklistTemplate} />
         <Route path="/history" component={ChecklistHistory} />
         <Route path="/settings" component={Settings} />
         <Route path="/templates" component={Templates} />
