+++ src/ChecklistDojo/ClientApp/src/components/ChecklistTemplate.css
