+++ src/ChecklistDojo/ClientApp/src/components/ChecklistTemplateItem.js
